👌 IMPROVE: Improved filter components
