+++ frontend/src/components/contact/Developer.jsx
@@ -5,7 +5,7 @@ const Container = styled.a`
   height: 4rem;
   width: 100%;
   margin: 0.5rem 0;
-  padding: 1rem 1.5rem;
+  padding: 0.5rem 1.5rem;
 
   display: flex;
   flex-direction: row;
