+++ frontend/src/components/courses/index.jsx
@@ -1,5 +1,6 @@
 export { default as CourseItem } from 'components/courses/item'
 export { default as CourseList } from 'components/courses/CourseList'
 export { default as CourseSearch } from 'components/courses/CourseSearch'
-export { default as FiltersBody } from 'components/courses/FiltersBody'
+export { default as Filters } from 'components/courses/Filters'
 export { default as PageNo } from 'components/courses/PageNo'
+export { default as FilterItem } from 'components/courses/FilterItem'
