+++ frontend/src/pages/Contact.jsx
@@ -2,7 +2,7 @@ import styled from 'styled-components'
 import { ContactForm, DeveloperList } from 'components/contact'
 
 const Container = styled.div`
-  margin: 4rem 0 0 11.5rem;
+  margin-left: 11.5rem;
 `
 
 const Contribute = () => {
